requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
datasets==3.5.1
dill==0.3.8
eval_type_backport==0.2.2
fastapi==0.115.12
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
geopandas==1.0.1
google-auth==2.39.0
google-genai==1.9.0
googlesearch-python==1.3.0
gradio==5.27.0
gradio_client==1.9.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.6.0
multidict==6.4.3
multiprocess==0.70.16
numpy==2.2.5
ollama==0.4.7
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyogrio==0.10.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruff==0.11.7
safehttpx==0.1.6
semantic-version==2.10.0
sgmllib3k==1.0.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
